added tests:
