--- colossalai/kernel/triton/llama_context_attention_kernel.py
@@ -1,9 +1,0 @@
-import torch
-try:
-    import triton
-    import triton.language as tl
-    HAS_TRITON = True
-except ImportError:
-    HAS_TRITON = False
-    print("please install triton from https://github.com/openai/triton")
-
