+++ tests/test_kernels/triton/test_self_attention_nonfusion.py
@@ -4,7 +4,7 @@ import torch
 from torch import nn 
 import torch.nn.functional as F
 
-from colossalai.kernel.triton.ops import self_attention_compute_using_triton
+from colossalai.kernel.triton.self_attention_nofusion import self_attention_compute_using_triton
 from colossalai.kernel.triton.qkv_matmul_kernel import qkv_gemm_4d_kernel
 
 try:
